+++ app/assets/javascripts/main.js
@@ -43,3 +43,5 @@
 //= require blueimp-gallery/blueimp-gallery-indicator
 //= require leaflet
 //= require api/authorization_page
+// = require bootstrap-markdown/bootstrap-markdown
+// = require helpers/markdown_editor