+++ features/support/publishing_cuke_helpers.rb
@@ -45,10 +45,8 @@ module PublishingCukeHelpers
   end
 
   def click_publisher
-    page.execute_script('
-     $("#publisher").removeClass("closed");
-     $("#publisher").find("#status_message_fake_text").focus();
-    ')
+    find("#status_message_fake_text").click
+    expect(find("#publisher")).to have_css(".publisher-textarea-wrapper.active")
   end
 
   def publisher_submittable?